Move ModelOptionFor function into the models package
Also rework UseRemoteRegistry client option.
